test(templating-router): Refactor tests
Literally none of the tests were working properly. They, for some reason, weren't throwing karma errors properly but a quick debug shows that they were all throwing browser errors. This commit completely rewrites the way tests work.

I was unable to understand the purpose of a few failing tests, which I have completely removed. Pull requests to reintroduce them are welcome.

This also includes some failing tests that are commented out that represent bugs that need to be resolved. They are useful tests, but they are not yet relevant.